Implement THREE.js
-Added use of three.js
-Management of persistant data changed
-Uses HTML5 method sessionStorage() to manage the storage of the parsed data
within the browser itself. This way it is persistant among all pages
-Added alert on the homepage to notify unsupported browsers that they need to update
-Added a LinkedList.js to help with data management
-Wrote manage3DWorld.js which enables a three.js world. Here I(mah985) tried my
best to set it up to be familiar to anyone who has used unity. Update and render
functions can be used as you would expect.
-Currently the manage3DWorld.js file doesnt do anything too exciting except construct
the same number of cubes in the world as there is parsed items from the csv file

-Controls coming next